+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - `Innmind\Http\Header\ContentLanguage` constructor is now private, use `::of()` named constructor
 - `Innmind\Http\Header\Link` constructor is now private, use `::of()` named constructor
 - `Innmind\Http\Header\WWWAuthenticate` constructor is now private, use `::of()` named constructor
+- `Innmind\Http\Header\CacheControl` constructor is now private, use `::of()` named constructor
 
 ### Removed